Add files via upload
This application will be able to model real life bank functions such as adding an account, increasing/decreasing the balance of an account, and allowing a user to check their balance. This application will demonstrate standard programming paradigms, and teach about essential programming concepts such as variables, commenting, input and output to the user with the Scanner class, and importing/using standard Java libraries.